+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/list/ListBuilder.java
@@ -18,7 +18,7 @@
  */
 package is.codion.swing.common.ui.component.list;
 
-import is.codion.common.value.ValueSet;
+import is.codion.common.value.Value;
 import is.codion.swing.common.ui.component.builder.ComponentBuilder;
 
 import javax.swing.JList;
@@ -29,83 +29,140 @@ import javax.swing.event.ListSelectionListener;
 import java.util.Set;
 
 /**
- * Builds a multi-selection JList.
+ * Builds a JList instance.
  * @param <T> the value type
+ * @param <V> the component value type
+ * @param <B> the builder type
+ * @see #factory(ListModel)
  */
-public interface ListBuilder<T> extends ComponentBuilder<Set<T>, JList<T>, ListBuilder<T>> {
+public interface ListBuilder<T, V, B extends ListBuilder<T, V, B>> extends ComponentBuilder<V, JList<T>, B> {
 
 	/**
 	 * @param visibleRowCount the visible row count
 	 * @return this builder instance
 	 * @see JList#setVisibleRowCount(int)
 	 */
-	ListBuilder<T> visibleRowCount(int visibleRowCount);
+	B visibleRowCount(int visibleRowCount);
 
 	/**
 	 * @param layoutOrientation the list layout orientation
 	 * @return thi builder instance
 	 * @see JList#setLayoutOrientation(int)
 	 */
-	ListBuilder<T> layoutOrientation(int layoutOrientation);
+	B layoutOrientation(int layoutOrientation);
 
 	/**
 	 * @param fixedCellHeight the fixed cell height
 	 * @return this builder instance
 	 * @see JList#setFixedCellHeight(int)
 	 */
-	ListBuilder<T> fixedCellHeight(int fixedCellHeight);
+	B fixedCellHeight(int fixedCellHeight);
 
 	/**
 	 * @param fixedCellWidth the fixed cell width
 	 * @return this builder instance
 	 * @see JList#setFixedCellWidth(int)
 	 */
-	ListBuilder<T> fixedCellWidth(int fixedCellWidth);
+	B fixedCellWidth(int fixedCellWidth);
 
 	/**
 	 * @param cellRenderer the cell renderer
 	 * @return this builder instance
 	 * @see JList#setCellRenderer(ListCellRenderer)
 	 */
-	ListBuilder<T> cellRenderer(ListCellRenderer<T> cellRenderer);
-
-	/**
-	 * @param selectionMode the list selection model
-	 * @return this builder instance
-	 * @see JList#setSelectionMode(int)
-	 */
-	ListBuilder<T> selectionMode(int selectionMode);
+	B cellRenderer(ListCellRenderer<T> cellRenderer);
 
 	/**
 	 * @param selectionModel the list selection model
 	 * @return this builder instance
 	 * @see JList#setSelectionModel(ListSelectionModel)
 	 */
-	ListBuilder<T> selectionModel(ListSelectionModel selectionModel);
+	B selectionModel(ListSelectionModel selectionModel);
 
 	/**
 	 * @param listSelectionListener the list selection listener
 	 * @return this builder instance
 	 * @see JList#addListSelectionListener(ListSelectionListener)
 	 */
-	ListBuilder<T> listSelectionListener(ListSelectionListener listSelectionListener);
+	B listSelectionListener(ListSelectionListener listSelectionListener);
 
 	/**
-	 * @param <T> the list element type
-	 * @param listModel the list model
-	 * @return a new builder instance
+	 * Builds a JList, where the value is represented by the list items.
+	 * @param <T> the value type
 	 */
-	static <T> ListBuilder<T> builder(ListModel<T> listModel) {
-		return new DefaultListBuilder<>(listModel, null);
+	interface Items<T> extends ListBuilder<T, Set<T>, Items<T>> {
+
+		/**
+		 * @param selectionMode the list selection model
+		 * @return this builder instance
+		 * @see JList#setSelectionMode(int)
+		 */
+		Items<T> selectionMode(int selectionMode);
 	}
 
 	/**
-	 * @param <T> the list element type
-	 * @param listModel the list model
-	 * @param linkedValueSet value set to link to the component
-	 * @return a new builder instance
+	 * Builds a multi-selection JList, where the value is represented by the selected items.
+	 * @param <T> the value type
 	 */
-	static <T> ListBuilder<T> builder(ListModel<T> listModel, ValueSet<T> linkedValueSet) {
-		return new DefaultListBuilder<>(listModel, linkedValueSet);
+	interface SelectedItems<T> extends ListBuilder<T, Set<T>, SelectedItems<T>> {}
+
+	/**
+	 * Builds a single-selection JList, where the value is represented by the selected item.
+	 * @param <T> the value type
+	 */
+	interface SelectedItem<T> extends ListBuilder<T, T, SelectedItem<T>> {}
+
+	/**
+	 * A factory for list builders, depending on what the component value should represent.
+	 */
+	interface Factory<T> {
+
+		/**
+		 * A JList builder, where the value is represented by the list items.
+		 * @return a JList builder
+		 */
+		Items<T> items();
+
+		/**
+		 * A JList builder, where the value is represented by the list items.
+		 * @param linkedValue the value to link to the list items
+		 * @return a JList builder
+		 */
+		Items<T> items(Value<Set<T>> linkedValue);
+
+		/**
+		 * A multi selection JList builder, where the value is represented by the selected items.
+		 * @return a JList builder
+		 */
+		SelectedItems<T> selectedItems();
+
+		/**
+		 * A multi selection JList builder, where the value is represented by the selected items.
+		 * @param linkedValue the value to link to the selected items
+		 * @return a JList builder
+		 */
+		SelectedItems<T> selectedItems(Value<Set<T>> linkedValue);
+
+		/**
+		 * A single-selection JList builder, where the value is represented by the selected item.
+		 * @return a JList builder
+		 */
+		SelectedItem<T> selectedItem();
+
+		/**
+		 * A single-selection JList builder, where the value is represented by the selected item.
+		 * @param linkedValue the value to link to the selected item
+		 * @return a JList builder
+		 */
+		SelectedItem<T> selectedItem(Value<T> linkedValue);
+	}
+
+	/**
+	 * @param listModel the list model to base the list on
+	 * @param <T> the list value type
+	 * @return a new list builder factory
+	 */
+	static <T> Factory<T> factory(ListModel<T> listModel) {
+		return new DefaultListBuilderFactory<>(listModel);
 	}
 }