+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntityComponents.java
@@ -33,6 +33,7 @@ import is.codion.swing.common.ui.component.button.CheckBoxBuilder;
 import is.codion.swing.common.ui.component.combobox.ComboBoxBuilder;
 import is.codion.swing.common.ui.component.combobox.ItemComboBoxBuilder;
 import is.codion.swing.common.ui.component.label.LabelBuilder;
+import is.codion.swing.common.ui.component.list.ListBuilder;
 import is.codion.swing.common.ui.component.slider.SliderBuilder;
 import is.codion.swing.common.ui.component.spinner.ItemSpinnerBuilder;
 import is.codion.swing.common.ui.component.spinner.ListSpinnerBuilder;
@@ -55,6 +56,7 @@ import javax.swing.JComboBox;
 import javax.swing.JComponent;
 import javax.swing.JTextField;
 import javax.swing.JToggleButton;
+import javax.swing.ListModel;
 import javax.swing.SpinnerListModel;
 import java.math.BigDecimal;
 import java.text.FieldPosition;
@@ -62,6 +64,7 @@ import java.text.Format;
 import java.text.ParsePosition;
 import java.time.temporal.Temporal;
 import java.util.Collection;
+import java.util.Set;
 import java.util.function.Supplier;
 
 import static is.codion.swing.common.model.component.combobox.ItemComboBoxModel.booleanItemComboBoxModel;
@@ -599,6 +602,43 @@ public class EntityComponents {
 						.toolTipText(attributeDefinition.description());
 	}
 
+	/**
+	 * @param listModel the list model to base the list on
+	 * @param <T> the value type
+	 * @return a builder factory for a list
+	 */
+	public final <T> EntityListBuilderFactory<T> list(ListModel<T> listModel) {
+		return new DefaultListBuilderFactory<>(listModel);
+	}
+
+	/**
+	 * A factory for list builders.
+	 * @param <T> the value type
+	 */
+	public interface EntityListBuilderFactory<T> {
+
+		/**
+		 * A JList builder, where the value is represented by the list items.
+		 * @param attribute the attribute
+		 * @return a JList builder
+		 */
+		ListBuilder.Items<T> items(Attribute<Set<T>> attribute);
+
+		/**
+		 * A multi selection JList builder, where the value is represented by the selected items.
+		 * @param attribute the attribute
+		 * @return a JList builder
+		 */
+		ListBuilder.SelectedItems<T> selectedItems(Attribute<Set<T>> attribute);
+
+		/**
+		 * A single selection JList builder, where the value is represented by the selected item.
+		 * @param attribute the attribute
+		 * @return a JList builder
+		 */
+		ListBuilder.SelectedItem<T> selectedItem(Attribute<T> attribute);
+	}
+
 	private static <T> FilteredComboBoxModel<T> createEnumComboBoxModel(Attribute<T> attribute, boolean nullable) {
 		FilteredComboBoxModel<T> comboBoxModel = new FilteredComboBoxModel<>();
 		Collection<T> enumConstants = asList(attribute.type().valueClass().getEnumConstants());
@@ -607,6 +647,39 @@ public class EntityComponents {
 		comboBoxModel.refresh();
 
 		return comboBoxModel;
+	}
+
+	private final class DefaultListBuilderFactory<T> implements EntityListBuilderFactory<T> {
+
+		private final ListBuilder.Factory<T> builderFactory;
+
+		private DefaultListBuilderFactory(ListModel<T> listModel) {
+			this.builderFactory = Components.list(listModel);
+		}
+
+		@Override
+		public ListBuilder.Items<T> items(Attribute<Set<T>> attribute) {
+			AttributeDefinition<Set<T>> attributeDefinition = entityDefinition.attributes().definition(attribute);
+
+			return builderFactory.items()
+							.toolTipText(attributeDefinition.description());
+		}
+
+		@Override
+		public ListBuilder.SelectedItems<T> selectedItems(Attribute<Set<T>> attribute) {
+			AttributeDefinition<Set<T>> attributeDefinition = entityDefinition.attributes().definition(attribute);
+
+			return builderFactory.selectedItems()
+							.toolTipText(attributeDefinition.description());
+		}
+
+		@Override
+		public ListBuilder.SelectedItem<T> selectedItem(Attribute<T> attribute) {
+			AttributeDefinition<T> attributeDefinition = entityDefinition.attributes().definition(attribute);
+
+			return builderFactory.selectedItem()
+							.toolTipText(attributeDefinition.description());
+		}
 	}
 
 	private static final class EntityReadOnlyFormat extends Format {